+++ ldpd/ldp_zebra.c
@@ -39,7 +39,6 @@ static void	 ifc2kaddr(struct interface *, struct connected *,
 		    struct kaddr *);
 static int	 ldp_zebra_send_mpls_labels(int, struct kroute *);
 static int	 ldp_router_id_update(ZAPI_CALLBACK_ARGS);
-static int	 ldp_interface_add(ZAPI_CALLBACK_ARGS);
 static int	 ldp_interface_delete(ZAPI_CALLBACK_ARGS);
 static int	 ldp_interface_status_change(ZAPI_CALLBACK_ARGS);
 static int	 ldp_interface_address_add(ZAPI_CALLBACK_ARGS);
@@ -264,19 +263,17 @@ ldp_router_id_update(ZAPI_CALLBACK_ARGS)
 }
 
 static int
-ldp_interface_add(ZAPI_CALLBACK_ARGS)
+ldp_ifp_create(struct interface *ifp)
 {
-	struct interface	*ifp;
 	struct kif		 kif;
 
-	ifp = zebra_interface_add_read(zclient->ibuf, vrf_id);
 	debug_zebra_in("interface add %s index %d mtu %d", ifp->name,
 	    ifp->ifindex, ifp->mtu);
 
 	ifp2kif(ifp, &kif);
 	main_imsg_compose_both(IMSG_IFSTATUS, &kif, sizeof(kif));
 
-	return (0);
+	return 0;
 }
 
 static int
@@ -532,11 +529,6 @@ ldp_zebra_connected(struct zclient *zclient)
 
 extern struct zebra_privs_t ldpd_privs;
 
-static int ldp_ifp_create(struct interface *ifp)
-{
-	return 0;
-}
-
 static int ldp_ifp_up(struct interface *ifp)
 {
 	return 0;
@@ -565,7 +557,6 @@ ldp_zebra_init(struct thread_master *master)
 	/* set callbacks */
 	zclient->zebra_connected = ldp_zebra_connected;
 	zclient->router_id_update = ldp_router_id_update;
-	zclient->interface_add = ldp_interface_add;
 	zclient->interface_delete = ldp_interface_delete;
 	zclient->interface_up = ldp_interface_status_change;
 	zclient->interface_down = ldp_interface_status_change;
