+++ nhrpd/nhrp_route.c
@@ -345,7 +345,6 @@ void nhrp_zebra_init(void)
 
 	zclient = zclient_new(master, &zclient_options_default);
 	zclient->zebra_connected = nhrp_zebra_connected;
-	zclient->interface_add = nhrp_interface_add;
 	zclient->interface_delete = nhrp_interface_delete;
 	zclient->interface_up = nhrp_interface_up;
 	zclient->interface_down = nhrp_interface_down;
