+++ ospfd/ospf_interface.h
@@ -321,7 +321,10 @@ extern int ospf_interface_neighbor_count(struct ospf_interface *oi);
    state of the interface. */
 extern void ospf_if_set_multicast(struct ospf_interface *);
 
+extern void ospf_if_interface(struct interface *ifp);
+
 DECLARE_HOOK(ospf_vl_add, (struct ospf_vl_data * vd), (vd))
 DECLARE_HOOK(ospf_vl_delete, (struct ospf_vl_data * vd), (vd))
 
+DECLARE_HOOK(ospf_if_update, (struct interface * ifp), (ifp))
 #endif /* _ZEBRA_OSPF_INTERFACE_H */
