+++ sharpd/sharp_zebra.c
@@ -58,15 +58,8 @@ static struct interface *zebra_interface_if_lookup(struct stream *s)
 }
 
 /* Inteface addition message from zebra. */
-static int interface_add(ZAPI_CALLBACK_ARGS)
+static int sharp_ifp_create(struct interface *ifp)
 {
-	struct interface *ifp;
-
-	ifp = zebra_interface_add_read(zclient->ibuf, vrf_id);
-
-	if (!ifp->info)
-		return 0;
-
 	return 0;
 }
 
@@ -392,11 +385,6 @@ static int sharp_nexthop_update(ZAPI_CALLBACK_ARGS)
 	return 0;
 }
 
-static int sharp_ifp_create(struct interface *ifp)
-{
-	return 0;
-}
-
 static int sharp_ifp_up(struct interface *ifp)
 {
 	return 0;
@@ -425,7 +413,6 @@ void sharp_zebra_init(void)
 
 	zclient_init(zclient, ZEBRA_ROUTE_SHARP, 0, &sharp_privs);
 	zclient->zebra_connected = zebra_connected;
-	zclient->interface_add = interface_add;
 	zclient->interface_delete = interface_delete;
 	zclient->interface_up = interface_state_up;
 	zclient->interface_down = interface_state_down;
