+++ vrrpd/vrrp_zebra.c
@@ -80,21 +80,9 @@ static int vrrp_router_id_update_zebra(int command, struct zclient *zclient,
 	return 0;
 }
 
-static int vrrp_zebra_if_add(int command, struct zclient *zclient,
-			     zebra_size_t length, vrf_id_t vrf_id)
+int vrrp_ifp_create(struct interface *ifp)
 {
-	struct interface *ifp;
-
-	/*
-	 * zebra api adds/dels interfaces using the same call
-	 * interface_add_read below, see comments in lib/zclient.c
-	 */
-	ifp = zebra_interface_add_read(zclient->ibuf, vrf_id);
-
-	if (!ifp)
-		return 0;
-
-	vrrp_zebra_debug_if_state(ifp, vrf_id, __func__);
+	vrrp_zebra_debug_if_state(ifp, ifp->vrf_id, __func__);
 
 	vrrp_if_add(ifp);
 
@@ -236,11 +224,6 @@ int vrrp_zclient_send_interface_protodown(struct interface *ifp, bool down)
 						down);
 }
 
-int vrrp_ifp_create(struct interface *ifp)
-{
-	return 0;
-}
-
 int vrrp_ifp_up(struct interface *ifp)
 {
 	return 0;
@@ -266,7 +249,6 @@ void vrrp_zebra_init(void)
 
 	zclient->zebra_connected = vrrp_zebra_connected;
 	zclient->router_id_update = vrrp_router_id_update_zebra;
-	zclient->interface_add = vrrp_zebra_if_add;
 	zclient->interface_delete = vrrp_zebra_if_del;
 	zclient->interface_up = vrrp_zebra_if_state_up;
 	zclient->interface_down = vrrp_zebra_if_state_down;
